+++ template/.nvim.lua
@@ -13,8 +13,12 @@ rust_analyzer.cargo.extraEnv = { RUSTUP_TOOLCHAIN = "esp" }
 rust_analyzer.server = { extraEnv = { RUSTUP_TOOLCHAIN = "stable" } }
 --ENDIF
 
+--IF option("neovim-rustaceanvim")
+vim.lsp.config("rust-analyzer", {
+--ELSE
 -- Note the neovim name of the language server is rust_analyzer with an underscore.
 vim.lsp.config("rust_analyzer", {
+--ENDIF
     settings = {
         ["rust-analyzer"] = rust_analyzer,
     },