+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update `esp-hal` dependency to ~1.1.0 (#334)
 - Reverted: Preserve per-method flashing options when switching probe-rs and espflash (#327)
 - Flashing options are now cleared when selecting a different flashing method (#327)