+++ CHANGELOG.md
@@ -266,8 +266,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 [Unreleased]: https://github.com/esp-rs/esp-generate/compare/v1.3.0...HEAD
-[1.3.0]: https://github.com/esp-rs/esp-generate/compare/v1.3.0-rc.0...v1.3.0
-[1.3.0-rc.0]: https://github.com/esp-rs/esp-generate/compare/v1.2.0...v1.3.0-rc.0
+[1.3.0]: https://github.com/esp-rs/esp-generate/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/esp-rs/esp-generate/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/esp-rs/esp-generate/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/esp-rs/esp-generate/compare/v1.0.0...v1.0.1